Extract the Featured Image DropZone to a separate component to make it reusable
