+++ client/post-editor/editor-drawer/featured-image.jsx
@@ -2,7 +2,6 @@
  * External dependencies
  */
 import React, { Component, PropTypes } from 'react';
-import { connect } from 'react-redux';
 import { localize } from 'i18n-calypso';
 import { head } from 'lodash';
 
@@ -15,11 +14,7 @@ import { getFeaturedImageId } from 'lib/posts/utils';
 import Accordion from 'components/accordion';
 import EditorDrawerWell from 'post-editor/editor-drawer-well';
 import FeaturedImage from 'post-editor/editor-featured-image';
-import MediaActions from 'lib/media/actions';
-import MediaUtils from 'lib/media/utils';
-
-import Dispatcher from 'dispatcher';
-import { editPost } from 'state/posts/actions';
+import FeaturedImageDropZone from 'post-editor/editor-featured-image/dropzone';
 
 class EditorDrawerFeaturedImage extends Component {
 	static propTypes = {
@@ -50,37 +45,13 @@ class EditorDrawerFeaturedImage extends Component {
 		stats.recordEvent( 'Featured image removed' );
 	}
 
-	onFilesDrop = ( files ) => {
-		/**
-		 * Filter files for `image` media prefix and return the first image.
-		 *
-		 * At the moment we ignore all the other images that were dragged onto the DropZone
-		 */
-		const droppedImage = head( MediaUtils.filterItemsByMimePrefix( files, 'image' ) );
-
-		if ( ! droppedImage ) {
-			return false;
-		}
-
-		Dispatcher.register( ( payload ) => {
-			const action = payload.action;
-
-			switch ( action.type ) {
-				case 'CREATE_MEDIA_ITEM':
-					// called when the transient blob has been created and the upload starts
-					break;
-				case 'RECEIVE_MEDIA_ITEM':
-					// called when the media file has been uploaded and needs to be refreshed from the server
-					setTimeout( () => {
-						PostActions.edit( {
-							featured_image: action.data.ID
-						} );
-					}, 0 );
-			}
-		} );
-
-		MediaActions.clearValidationErrors( this.props.site.ID );
-		MediaActions.add( this.props.site.ID, [ droppedImage ] );
+	getDropZone = () => {
+		return (
+			<FeaturedImageDropZone
+				site={ this.props.site }
+				post={ this.props.post }
+			/>
+		);
 	};
 
 	render() {
@@ -92,7 +63,7 @@ class EditorDrawerFeaturedImage extends Component {
 					label={ translate( 'Set Featured Image' ) }
 					empty={ ! site || ! post || ! getFeaturedImageId( post ) }
 					onClick={ this.startSelecting }
-					onFileDrop={ this.onFilesDrop }
+					customDropZone={ this.getDropZone() }
 					onRemove={ this.removeImage }
 				>
 					<FeaturedImage
@@ -107,6 +78,4 @@ class EditorDrawerFeaturedImage extends Component {
 	}
 }
 
-export default connect( null, {
-	editPost
-} )( localize( EditorDrawerFeaturedImage ) );
+export default localize( EditorDrawerFeaturedImage );
